updated algorithm to handle 2 sequential bad phases
